cleanup example files
